Change VanillaVAE decoder activation to sigmoid

+++ multimodal_dataset_bias_toy/vae/experiment.py
@@ -39,7 +39,7 @@ if __name__ == "__main__":
     parser.add_argument("--lr", type=float, default=1e-4)
     parser.add_argument("--batch_size", type=int, default=128)
     parser.add_argument("--n_epochs", type=int, default=1000)
-    parser.add_argument("--early_stop_ratio", type=int, default=0.1)
+    parser.add_argument("--early_stop_ratio", type=float, default=0.1)
     parser.add_argument("--n_gpus", type=int, default=1)
     parser.add_argument("--n_workers", type=int, default=20)
     args = parser.parse_args()
